requirements.txt
numpy
pandas
datasets
astropy
dl
h5py
aiohttp
kcorrect
matplotlib
tqdm
jupyter
kaggle
seaborn
scikit-learn
astro-datalab
lightning[extra]
torch==2.0.0+cu117
dinov2 @ git+https://github.com/facebookresearch/dinov2.git
torchmetrics==0.10.3
torchvision==0.15.0
scikit-image
tables
pyro-ppl
